Organize tests into unit and integration test suites.
- Move unit tests (AfterDueByTest, AfterExistingForTest, CompletionAwareTest, UtilsTest) to tests/Unit/
- Move integration tests (BoundaryConditionTest, CarbonTestHelpersTest, ComprehensiveEdgeCaseTest, etc.) to tests/Integration/
- Update namespaces for all moved test files
- Configure PHPUnit with separate test suites for unit and integration tests
- Add composer scripts: test:unit and test:integration for easier test execution
