+++ composer.json
@@ -37,7 +37,9 @@
         "format": "php-cs-fixer fix",
         "format:check": "php-cs-fixer fix --dry-run --diff",
         "test": "phpunit",
-        "test:coverage": "XDEBUG_MODE=coverage phpunit --configuration phpunit-coverage.xml"
+        "test:coverage": "XDEBUG_MODE=coverage phpunit --configuration phpunit-coverage.xml",
+        "test:unit": "phpunit --testsuite unit",
+        "test:integration": "phpunit --testsuite integration"
     },
     "config": {
         "optimize-autoloader": true,
